phase-3a: A6 — a review's rating is a half-step, 1 to 5: parsed.REVIEW_RATINGS declares the closed set once, review_rating parses a digit or a digit and .5 strictly, the Opinion Assurances parser refuses anything else naming the value, load_reviews checks every row against the set and inserts nothing on a refusal, raw_reviews.rating is decimal(2,1), content_hash spells a decimal one way so the synthetic hashes do not move; the sample re-frozen with one half-step (Freeze: fixtures/opinion-assurances/); pinned by each half-step accepted, off-set values refused, the loaded row read back as parsed, the hash of 1, '1' and 1.0 equal; restores invariant 6 — a page in the declared shape loads, and the shape is the site's (first live run, 2026-09-03)

+++ sql/raw/raw_reviews.sql
@@ -2,6 +2,8 @@
 -- Grain: one row per capture; natural key (source, external_id, content_hash).
 --   A re-capture of an unchanged review inserts nothing; an edited review has a
 --   new content_hash and appends a new row, so history is kept.
+-- rating: a half-step 1 to 5 (ingest/parsed.py::REVIEW_RATINGS, checked at the
+--   load); the profile site rates in half stars, the app feed in digits (A6).
 -- Provenance: source, source_url, captured_at, run_id. captured_at is the
 --   capture time carried in from the fixture/scraper, run_id is stamped by the
 --   loader in Python — no clock in SQL.
@@ -14,7 +16,7 @@ create table if not exists raw_reviews (
     captured_at   text    not null,
     run_id        text    not null,
     review_date   text    not null,
-    rating        integer not null,
+    rating        decimal(2, 1) not null,
     title         text    not null,
     body          text    not null,
     content_hash  text    not null
